+++ util/dmdeleter.py
@@ -3,10 +3,9 @@
 # Hazard Nuker under the GNU General Public Liscense v2 (1991).
 
 import requests
-import Hazard
 from colorama import Fore
 
-from util.plugins.common import print_slow, getheaders, proxy
+from util.plugins.common import getheaders, proxy
 
 def DmDeleter(token, channels):
     for channel in channels: