+++ util/login.py
@@ -6,10 +6,10 @@ import requests
 import Hazard
 
 from time import sleep
-from selenium import webdriver
+from selenium import webdriver, common
 from colorama import Fore, Back
 
-from util.plugins.common import get_driver, getheaders
+from util.plugins.common import getDriver, getheaders, SlowPrint
 
 def TokenLogin(token):
     j = requests.get("https://discord.com/api/v9/users/@me", headers=getheaders(token)).json()
@@ -18,32 +18,53 @@ def TokenLogin(token):
             document.body.appendChild(document.createElement `iframe`).contentWindow.localStorage.token = `"%s"`
             location.reload();
         """ % (token)
-    type_ = get_driver()
+    type_ = getDriver()
 
     if type_ == "chromedriver.exe":
         opts = webdriver.ChromeOptions()
         opts.add_experimental_option('excludeSwitches', ['enable-logging'])
         opts.add_experimental_option("detach", True)
-        driver = webdriver.Chrome(options=opts)
-    # elif type_ == "operadriver.exe":
-    #     opts = webdriver.opera.options.ChromeOptions()
-    #     opts.add_experimental_option('excludeSwitches', ['enable-logging'])
-    #     opts.add_experimental_option("detach", True)
-    #     driver = webdriver.Opera(options=opts)
+        try:
+            driver = webdriver.Chrome(options=opts)
+        except common.exceptions.SessionNotCreatedException as e:
+            print(e.msg)
+            sleep(2)
+            SlowPrint("Enter anything to continue. . . ")
+            input()
+            Hazard.main()
+    elif type_ == "operadriver.exe":
+        opts = webdriver.opera.options.ChromeOptions()
+        opts.add_experimental_option('excludeSwitches', ['enable-logging'])
+        opts.add_experimental_option("detach", True)
+        try:
+            driver = webdriver.Opera(options=opts)
+        except common.exceptions.SessionNotCreatedException as e:
+            print(e.msg)
+            sleep(2)
+            SlowPrint("Enter anything to continue. . . ")
+            input()
+            Hazard.main()
     elif type_ == "msedgedriver.exe":
         opts = webdriver.EdgeOptions()
         opts.add_experimental_option('excludeSwitches', ['enable-logging'])
         opts.add_experimental_option("detach", True)
-        driver = webdriver.Edge(options=opts)
+        try:
+            driver = webdriver.Edge(options=opts)
+        except common.exceptions.SessionNotCreatedException as e:
+            print(e.msg)
+            sleep(2)
+            SlowPrint(f"Enter anything to continue. . .")
+            input()
+            Hazard.main()
     else:
-        print(f'{Fore.RESET}[{Fore.RED}Error{Fore.RESET}] : Coudln\'t find a driver to automate the proccess of login in to {user}')
+        print(f'{Fore.RESET}[{Fore.RED}Error{Fore.RESET}] : Coudln\'t find a suitable driver to automatically login to {user}')
         sleep(3)
         print(f"{Fore.YELLOW}Paste this script into the console of a browser:{Fore.RESET}\n\n{Back.RED}{script}\n{Back.RESET}")
         print("Enter anything to continue. . . ", end="")
         input()
         Hazard.main()
 
-    print(f"{Fore.GREEN}Logging into {Fore.BLUE}{user}")
+    print(f"{Fore.GREEN}Logging into {Fore.CYAN}{user}")
     driver.get("https://discordapp.com/login")
     driver.execute_script(script)
     Hazard.main()