+++ util/profilechanger.py
@@ -6,7 +6,7 @@ import requests
 import Hazard
 from colorama import Fore
 
-from util.plugins.common import print_slow, getheaders, proxy
+from util.plugins.common import SlowPrint, getheaders, proxy
 
 def HouseChanger(token, _type):
     house = {
@@ -17,7 +17,7 @@ def HouseChanger(token, _type):
     #change hypesquad
     hypesqad_req = {'house_id': _type}
     requests.post('https://discord.com/api/v9/hypesquad/online', headers=getheaders(token), json=hypesqad_req)
-    print_slow(f"\n{Fore.GREEN}Hypesquad changed to {Fore.WHITE}{house[_type]}{Fore.GREEN} ")
+    SlowPrint(f"\n{Fore.GREEN}Hypesquad changed to {Fore.WHITE}{house[_type]}{Fore.GREEN} ")
     print("Enter anything to continue. . . ", end="")
     input()
     Hazard.main()
@@ -27,7 +27,7 @@ def StatusChanger(token, Status):
     custom_status = {"custom_status": {"text": Status}} #{"text": Status, "emoji_name": "☢"} if you want to add an emoji to the status
     try:
         requests.patch("https://discord.com/api/v9/users/@me/settings", proxies={"http": f'{proxy()}'}, headers=getheaders(token), json=custom_status)
-        print_slow(f"\n{Fore.GREEN}Status changed to {Fore.WHITE}{Status}{Fore.GREEN} ")
+        SlowPrint(f"\n{Fore.GREEN}Status changed to {Fore.WHITE}{Status}{Fore.GREEN} ")
     except Exception as e:
         print(f"{Fore.RED}Error:\n{e}\nOccurred while trying to change the status :/")
     print("Enter anything to continue. . . ", end="")
@@ -39,7 +39,7 @@ def BioChanger(token, bio):
     custom_bio = {"bio": str(bio)}
     try:
         requests.patch("https://discord.com/api/v9/users/@me", proxies={"http": f'{proxy()}'}, headers=getheaders(token), json=custom_bio)
-        print_slow(f"\n{Fore.GREEN}Bio changed to {Fore.WHITE}{bio}{Fore.GREEN} ")
+        SlowPrint(f"\n{Fore.GREEN}Bio changed to {Fore.WHITE}{bio}{Fore.GREEN} ")
     except Exception as e:
         print(f"{Fore.RED}Error:\n{e}\nOccurred while trying to change the status :/")
     print("Enter anything to continue. . . ", end="")